updating suggested questions
