+++ llm-service/app/routers/index/__init__.py
@@ -39,12 +39,7 @@
 import logging
 
 from fastapi import APIRouter
-from pydantic import BaseModel
 
-from ... import exceptions
-from ...services import qdrant, s3
-from ...services.chat import (generate_suggested_questions, v2_chat)
-from ...services.chat_store import RagContext, RagStudioChatMessage
 from . import data_source
 from . import sessions
 from . import amp_update
@@ -62,38 +57,5 @@ router.include_router(amp_update.router, prefix="/index", deprecated=True)
 router.include_router(models.router)
 
 
-class SuggestQuestionsRequest(BaseModel):
-    data_source_id: int
-    chat_history: list[RagContext]
-    configuration: qdrant.RagPredictConfiguration = qdrant.RagPredictConfiguration()
-
-class RagSuggestedQuestionsResponse(BaseModel):
-    suggested_questions: list[str]
 
 
-
-@router.post("/suggest-questions", summary="Suggest questions with context")
-@exceptions.propagates
-def suggest_questions(
-    request: SuggestQuestionsRequest,
-) -> RagSuggestedQuestionsResponse:
-    data_source_size = qdrant.size_of(request.data_source_id)
-    qdrant.check_data_source_exists(data_source_size)
-    suggested_questions = generate_suggested_questions(
-        request.configuration, request.data_source_id, request.chat_history, data_source_size
-    )
-    return RagSuggestedQuestionsResponse(suggested_questions=suggested_questions)
-
-class RagStudioChatRequest(BaseModel):
-    data_source_id: int
-    session_id: int
-    query: str
-    configuration: qdrant.RagPredictConfiguration
-
-
-@router.post("/chat", summary="Chat with your documents in the requested datasource")
-@exceptions.propagates
-def chat(
-    request: RagStudioChatRequest,
-) -> RagStudioChatMessage:
-    return v2_chat(request.session_id, request.data_source_id, request.query, request.configuration)
