+++ llm-service/app/routers/index/sessions/__init__.py
@@ -38,8 +38,12 @@
 
 from fastapi import APIRouter
 
+from pydantic import BaseModel
+
 from .... import exceptions
-from ....services.chat_store import RagStudioChatMessage, chat_store
+from ....services.chat_store import RagStudioChatMessage, chat_store, RagContext
+from ....services import qdrant
+from ....services.chat import (v2_chat, generate_suggested_questions)
 
 router = APIRouter(prefix="/sessions/{session_id}", tags=["Sessions"])
 
@@ -60,3 +64,39 @@ def delete_chat_history(session_id: int) -> str:
     chat_store.delete_chat_history(session_id=session_id)
     return "Chat history deleted."
 
+
+class RagStudioChatRequest(BaseModel):
+    data_source_id: int
+    query: str
+    configuration: qdrant.RagPredictConfiguration
+
+
+@router.post("/chat", summary="Chat with your documents in the requested datasource")
+@exceptions.propagates
+def chat(
+        session_id: int,
+        request: RagStudioChatRequest,
+) -> RagStudioChatMessage:
+    return v2_chat(session_id, request.data_source_id, request.query, request.configuration)
+
+
+class SuggestQuestionsRequest(BaseModel):
+    data_source_id: int
+    chat_history: list[RagContext]
+    configuration: qdrant.RagPredictConfiguration = qdrant.RagPredictConfiguration()
+
+class RagSuggestedQuestionsResponse(BaseModel):
+    suggested_questions: list[str]
+
+@router.post("/suggest-questions", summary="Suggest questions with context")
+@exceptions.propagates
+def suggest_questions(
+        session_id: int,
+        request: SuggestQuestionsRequest,
+) -> RagSuggestedQuestionsResponse:
+    data_source_size = qdrant.size_of(request.data_source_id)
+    qdrant.check_data_source_exists(data_source_size)
+    suggested_questions = generate_suggested_questions(
+        request.configuration, request.data_source_id, data_source_size, session_id
+    )
+    return RagSuggestedQuestionsResponse(suggested_questions=suggested_questions)
